requirements.txt
Faker
apache-airflow-providers-postgres
matplotlib
